+++ parqueteur/dump.go
@@ -41,8 +41,8 @@ func runDump(cmd *Command, args []string) error {
 
 	const batch = 16
 	values := make([]interface{}, batch, batch)
-	dLevels := make([]int, batch, batch)
-	rLevels := make([]int, batch, batch)
+	dLevels := make([]uint16, batch, batch)
+	rLevels := make([]uint16, batch, batch)
 	var n int
 	for rg, _ := range f.MetaData.RowGroups {
 		cr, err := f.NewReader(col, rg)
